fix: move changes to maintained directory (per comment)
