+++ packages/mui-base/src/useAutocomplete/useAutocomplete.spec.ts
@@ -1,9 +1,5 @@
 import { expectType } from '@mui/types';
-import {
-  useAutocomplete,
-  FilterOptionsState,
-  AutocompleteGroupedOption,
-} from '@mui/base/useAutocomplete';
+import { useAutocomplete, FilterOptionsState } from '@mui/base/useAutocomplete';
 
 interface Person {
   id: string;
@@ -185,17 +181,4 @@ function Component() {
     },
     freeSolo: true,
   });
-
-  const ungroupedAutocomplete = useAutocomplete({ options: persons });
-  expectType<Person[], typeof ungroupedAutocomplete.groupedOptions>(
-    ungroupedAutocomplete.groupedOptions,
-  );
-
-  const groupedAutocomplete = useAutocomplete({
-    options: persons,
-    groupBy: ({ id }) => id,
-  });
-  expectType<AutocompleteGroupedOption<Person>[], typeof groupedAutocomplete.groupedOptions>(
-    groupedAutocomplete.groupedOptions,
-  );
 }
